% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{glioma_deconv}
\alias{glioma_deconv}
\title{Glioma deconv}
\usage{
glioma_deconv(exp)
}
\arguments{
\item{exp}{the expression matrix, with no log scale}
}
\description{
This function generatea a deconvolution result of bulk RNA-seq
}
\examples{
res <- glioma_deconv(exp)
}
