% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_barplot.R
\name{vis_barplot}
\alias{vis_barplot}
\title{Glioma deconv}
\usage{
vis_barplot(res, use_phe = F, phe = NULL)
}
\arguments{
\item{res}{the deconvolution result}

\item{use_phe}{if using phe, default is False}

\item{phe}{phe, rownames should be in rownames of deconvolution result}
}
\description{
This function generatea a barplot of deconvolution result
}
\examples{
vis_barplot(res)

}
