% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_boxplot.R
\name{vis_boxplot}
\alias{vis_boxplot}
\title{Glioma deconv}
\usage{
vis_boxplot(res, phe, group, celltype = "all", ncol = 5)
}
\arguments{
\item{res}{the deconvolution result}

\item{phe}{phe, rownames should be in rownames of deconvolution result}

\item{group}{group name, should be in phe}

\item{celltype}{default is "all", can be "NPClike", "AClike", "OPClike","MESlike","Microglia","Granulocyte","Dendritic cells","MHC.TAM", "Ribo.TAM", "Prolife.TAM","Lipid.TAM", "Hypoxia.TAM","Senescent.TAM","Oligodendrocyte","Fibroblast","Lymphocyte", "Endothelial"}

\item{ncol}{n columns of the plot}
}
\description{
This function generatea a boxplot of deconvolution result
}
\examples{
vis_boxplot(res)
}
